+++ src/System/FSNotify/Listener.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE Rank2Types #-}
 --
 -- Copyright (c) 2012 Mark Dittmer - http://www.markdittmer.org
 -- Developed for a Google Summer of Code project - http://gsoc2012.markdittmer.org
@@ -8,6 +11,7 @@ module System.FSNotify.Listener
        , epsilonDefault
        , FileListener(..)
        , StopListening
+       , ListenFn
        , newDebouncePayload
        ) where
 
@@ -22,11 +26,13 @@ import System.FilePath
 -- | An action that cancels a watching/listening job
 type StopListening = IO ()
 
+type ListenFn sessionType argType = FileListener sessionType argType => WatchConfig -> sessionType -> FilePath -> ActionPredicate -> EventCallback -> IO StopListening
+
 -- | A typeclass that imposes structure on watch managers capable of listening
 -- for events, or simulated listening for events.
-class FileListener sessionType where
+class FileListener sessionType argType | sessionType -> argType where
   -- | Initialize a file listener instance.
-  initSession :: IO (Either Text sessionType)
+  initSession :: Int -> IO (Either Text sessionType)
   -- ^ An initialized file listener, or a reason why one wasn't able to start.
 
   -- | Kill a file listener instance.
@@ -38,13 +44,13 @@ class FileListener sessionType where
   -- Listening for events associated with immediate contents of a directory will
   -- only report events associated with files within the specified directory, and
   -- not files within its subdirectories.
-  listen :: WatchConfig -> sessionType -> FilePath -> ActionPredicate -> EventChannel -> IO StopListening
+  listen :: ListenFn sessionType argType
 
   -- | Listen for file events associated with all the contents of a directory.
   -- Listening for events associated with all the contents of a directory will
   -- report events associated with files within the specified directory and its
   -- subdirectories.
-  listenRecursive :: WatchConfig -> sessionType -> FilePath -> ActionPredicate -> EventChannel -> IO StopListening
+  listenRecursive :: ListenFn sessionType argType
 
   -- | Does this manager use polling?
   usesPolling :: sessionType -> Bool