+++ CHANGELOG.md
@@ -6,7 +6,11 @@ Version 0.4.0.0
 
 API breaking update.
 
-* New `confThreadPerEvent` option in `WatchConfig`
+* New options for threading control (single-threaded, thread-per-watch, and thread-per-manager)
+* Revamp `WatchConfig` options to be less confusing reduce boolean blindness.
+* Pull out debouncing stuff, since it was never correct.
+* Don't silently fall back to polling on failure of native watcher.
+  Instead, throw an exception which the user can recover from by switching to polling.
 
 Version 0.3.0.0
 ---------------